requirements.txt
pandas
tavily-python
